+++ postgres/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ***Added***:
 
+* Add schema collection to Postgres integration ([#15484](https://github.com/DataDog/integrations-core/pull/15484))
 * Add support for sending `database_instance` metadata ([#15559](https://github.com/DataDog/integrations-core/pull/15559))
 * Update dependencies for Agent 7.48 ([#15585](https://github.com/DataDog/integrations-core/pull/15585))